backend/app/requirements.txt
# Database
mysql-connector-python==9.0.0
redis==5.0.8
SQLAlchemy==2.0.32

# Service
fastapi==0.112.0
gunicorn==22.0.0
passlib==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
email_validator==2.2.0
PyJWT==2.9.0
passlib[bcrypt]
pyyaml==6.0.2
python-dotenv==1.0.1
starlette==0.37.2
uvicorn==0.30.5
pymilvus
python-multipart
faiss-cpu
pymilvus
openai
langgraph
langchain 
langchain-openai
scikit-learn
flask
feedparser